exo1/requirements.txt
numpy
matplotlib.pyplot
datetime